+++ src/main/java/fr/alasdiablo/janoeo/world/gen/NetherOreGenerator.java
@@ -1,5 +1,6 @@
 package fr.alasdiablo.janoeo.world.gen;
 
+import fr.alasdiablo.janoeo.block.NetherOre;
 import fr.alasdiablo.janoeo.init.NetherOresBlocks;
 import fr.alasdiablo.janoeo.config.GlobalConfig;
 import fr.alasdiablo.janoeo.config.NetherConfig;
@@ -12,8 +13,6 @@ public class NetherOreGenerator implements IWorldGenerator {
     public void startWorldGeneration(Biome biome) {
         GlobalConfig.Config globalConfig = GlobalConfig.CONFIG;
         NetherConfig.Config netherConfig = NetherConfig.CONFIG;
-
-
 
         if (globalConfig.NETHER_ORE_GEN.get()) {
             if (netherConfig.COAL_NETHER_ORE.get()) {
@@ -45,6 +44,50 @@ public class NetherOreGenerator implements IWorldGenerator {
                         biome,
                         WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.LAPIS_NETHER_ORE.getRegistryName())
                 );
+            }
+            if (globalConfig.EXTRA_NETHER_ORE_GEN.get()) {
+                if (netherConfig.COPPER_NETHER_ORE.get()) {
+                    OreGenUtils.addFeatureToBiome(
+                            biome,
+                            WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.COPPER_NETHER_ORE.getRegistryName())
+                    );
+                }
+                if (netherConfig.ALUMINIUM_NETHER_ORE.get()) {
+                    OreGenUtils.addFeatureToBiome(
+                            biome,
+                            WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.ALUMINIUM_NETHER_ORE.getRegistryName())
+                    );
+                }
+                if (netherConfig.RUBY_NETHER_ORE.get()) {
+                    OreGenUtils.addFeatureToBiome(
+                            biome,
+                            WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.RUBY_NETHER_ORE.getRegistryName())
+                    );
+                }
+                if (netherConfig.SAPPHIRE_NETHER_ORE.get()) {
+                    OreGenUtils.addFeatureToBiome(
+                            biome,
+                            WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.SAPPHIRE_NETHER_ORE.getRegistryName())
+                    );
+                }
+                if (netherConfig.SILVER_NETHER_ORE.get()) {
+                    OreGenUtils.addFeatureToBiome(
+                            biome,
+                            WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.SILVER_NETHER_ORE.getRegistryName())
+                    );
+                }
+                if (netherConfig.TIN_NETHER_ORE.get()) {
+                    OreGenUtils.addFeatureToBiome(
+                            biome,
+                            WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.TIN_NETHER_ORE.getRegistryName())
+                    );
+                }
+                if (netherConfig.URANIUM_NETHER_ORE.get()) {
+                    OreGenUtils.addFeatureToBiome(
+                            biome,
+                            WorldGenRegistries.field_243653_e.getOrDefault(NetherOresBlocks.URANIUM_NETHER_ORE.getRegistryName())
+                    );
+                }
             }
         }
     }