+++ changelog.md
@@ -1,6 +1,9 @@
+### 4.4.0.53-experimental
++ add nether world generation
+
 ### 4.4.0.52-experimental
-Testing version of the port to 1.16.3
-Only overworld ore generation work
++ Testing version of the port to 1.16.3
+    + Only overworld ore generation work
 
 ### 4.4.0.49-experimental
 + drop support for 1.16 for now (wait stable release)